pyproject.toml
[project]
name = "spannergraphuser360chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dataclass-csv>=1.4.0",
    "google-adk==1.2.1",
    "google-cloud-spanner>=3.58.0",
    "langchain>=0.3.27",
    "toolbox-core>=0.5.2",
    "toolbox-langchain>=0.5.2",
]
